Moved debug menus to separate directory
