+++ Client/Main.cs
@@ -1,6 +1,7 @@
 ﻿using CitizenFX.Core;
 using CitizenFX.Core.Native;
 using PocceMod.Client.Menus;
+using PocceMod.Client.Menus.Dev;
 using PocceMod.Shared;
 using System;
 using System.Collections.Generic;
